Removed player manager and threads from bukkit plugin
