+++ bukkit/src/main/java/com/noahhusby/sledgehammer/server/util/SkullUtil.java
@@ -191,8 +191,6 @@ public class SkullUtil {
             }
             metaSetProfileMethod.invoke(meta, makeProfile(b64));
         } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException ex) {
-            // if in an older API where there is no setProfile method,
-            // we set the profile field directly.
             try {
                 if (metaProfileField == null) {
                     metaProfileField = meta.getClass().getDeclaredField("profile");
